initial can changes
